refactor: re-arrange unit tests
